vec -> Bytes + comments
